fix lib refactor
